Draw Time-Heatmap Using enrichR

This funcitons takes the master.list output from run_TrendCatcher. And apply a time window sliding strategy to capture all the genes increased/decreased compared to its previous break point, and apply enrichR enrichment analysis.

draw_TimeHeatmap_enrichR(
  master.list,
  logFC.thres = 1,
  top.n = 10,
  dyn.gene.p.thres = 0.05,
  dbs = "BioPlanet_2019",
  term.width = 80,
  OrgDb = "org.Mm.eg.db",
  GO.enrich.p = 0.05,
  figure.title = "",
  save.tiff.path = NA,
  tiff.res = 100,
  tiff.width = 1500,
  tiff.height = 1500
)

Arguments

master.list,

a list object. The output from run_TrendCatcher function, contains master.table element.

logFC.thres,

a numeric variable. The logFC threshold compared to each genes previous break point expression level. By default is 1, meaning for each gene, the current time window's expression level is 2-fold compared to previous break point's expression level.

top.n,

an integer variable. The top N GO enrichment term need to be shown in the Time-Heatmap for up and down regulated pathway. By default is 10. Top 20 GO terms, 10 from up-regulated pathway and 10 from down-regulated pathway will shown in Time-Heatmap.

dyn.gene.p.thres,

a numeric variable. The DDEGs dynamic p-value threshold. By default is 0.05.

dbs,

must one of the enrichR supported database name. To check the list, run dbs <- listEnrichrDbs() command. By default is "BioPlanet_2019".

term.width,

an integer variable. The character length for each GO term. If one GO term is super long, we can wrap it into term.width of strings into multiple rows. By default if 80.

GO.enrich.p,

an numeric variable. The GO enrichment p-value threshold. By default if 0.05.

figure.title,

a character variable. The main title of Time-Heatmap.

id.ensembl,

a logic variable. If using ensembl as the keytype. This must match the row name of master.table. By default is TRUE.

ont,

one of "BP", "MF", and "CC" subontologies, or "ALL" for all three. By default is "BP".

save.pdf.path,

a character variable. If need to save the figure into PDF file. This must be an absolute file path. If not needed save as PDF file, set it to NA. By defualt iS NA.

pdf.width,

a numeric variable. The width of PDF file. By default is 13.

pdf.height,

a numeric variable. The height of PDF file. By default is 15.

Value

A list object, including elements names merge.df and time.heatmap. time.heatmap is the ggplot object. merge.df includes all the enrichR enrichment result and activation/deactivation time.

Examples

if (FALSE) {
example.file.path<-system.file("extdata", "BrainMasterList.rda", package = "TrendCatcher")
load(example.file.path)
th.obj<-draw_TimeHeatmap_enrichR(master.list = master.list)
print(th.obj$time.heatmap)
head(th.obj$merge.df)
}