Preprocessing for TrendCatcher

This is the preprocessing function to prepare the input count table for run_TrendCatcher function. It takes the CSV file count table and logic variables to check if normalization and batch correction needed for preprocessing. It creates pdf figure report to show before and after of normalization and batch correction to assess quality control (QC).

preprocess_TrendCatcher(
  count.table.path = "",
  need.batch.correction = TRUE,
  need.normalization = TRUE,
  batch.arr = "",
  pdf.file.path = NA,
  pdf.width = 8,
  pdf.height = 10,
  n.low.count = 10
)

Arguments

count.table.path,

string contain the absolute path of the CSV file count table, with first column as GENE SYMBOL or GENE ENSEMBL and first row as SAMPLE NAME (with format composed by project name,time and replicateID, such as "Lung_0_Rep1")

need.batch.correction,

logic variable. If batch correction is needed. By default is TRUE.

need.normalization,

logic variable. If normalization is needed. By default is TRUE.

batch.arr,

a numeric vector of batch number. Need to be the same length as the number of samples.

pdf.file.path,

an absolute file path for save the QC report file. If not need, set if to NA. The report will be printed. By default is NA.

pdf.width,

a numeric variable. The PDF file width size. By default is 8.

pdf.height,

a numeric variable. The PDF file height size. By default is 10.

n.low.count,

a numeric variable. The minimal number to filter low count genes. By default is 10.

Value

A matrix array object.

Examples

example.file.path<-system.file("extdata", "Brain_DemoRawCountTable.csv", package = "TrendCatcher")
if (FALSE) {
count.table<-preprocess_TrendCatcher(count.table.path = example.file.path,
need.batch.correction = TRUE,
need.normalization = TRUE,
batch.arr<-c(2,2,3,3,3,
             3,3,3,3,0,
             2,3,0,4,
             3,3,2,2,0,4,
             3,3,2,0,
              2,2,0,4),
pdf.file.path = NA,
pdf.width=8, pdf.height=10,
n.low.count = 10)
}