Run TrendCatcher Main Algorithm

This is the main function to run TrendCatcher to identify Dynamic Differentially Expressed Genes (DDEGs). This function loads a rounded count matrix CSV file after the normalization and batch correction, run the core algorithm and output a list object contains all the genes dynamic information.

run_TrendCatcher(
  count.table.path = "~/Documents/TrendCatcher/inst/extdata/Lung_DemoCountTable.csv",
  baseline.t = 0,
  time.unit = "h",
  min.low.count = 1,
  para.core.n = NA,
  dyn.p.thres = 0.05,
  show.verbose = F
)

Arguments

count.table.path,

string contain the absolute path of the CSV file count table, with first column as GENE SYMBOL or GENE ENSEMBL and first row as SAMPLE NAME (with format composed by project name,time and replicateID, such as "Lung_0_Rep1")

baseline.t,

one numeric variable, the baseline time of the longitudinal study. By default it is 0.

time.unit,

one character variable, the time unit of longitudinal study. If choose hour, please transform all sample collecting time into hour.

min.low.count,

one numeric variable, the minimal count threshold for filtering low count within each time group. By default it is 1.

para.core.n,

one numeric variable, number of cores will be used for running TrendCatcher parallel. By default it is NA, which will use N-1 cores from computer.

dyn.p.thres,

one numeric variable, the threshold of adjusted p-value of the dynamic gene. By default 0.05.

show.verbose,

logic variable. If gssanova fitting failed, users can set this to TRUE, it will print out which gene failed the fitting. This process takes only one CPU, so it may be slower than the multi-core version. Normally the fitting failure is caused by low count genes. Users can manually remove it from your count table. By default set to FALSE.

Value

A list object, including "time.unit", "baseline.t", "t.arr", "Project.name", "raw.df", "fitted.count" and "master.table".

Examples

example.file.path<-system.file("extdata", "Brain_DemoCountTable.csv", package = "TrendCatcher")
if (FALSE) {
master.list<-run_TrendCatcher(count.table.path = example.file.path,
baseline.t = 0,
time.unit = "h",
min.low.count = 1,
para.core.n = NA,
dyn.p.thres = 0.05,
show.verbose = FALSE)
}